requirements.txt
gradio
pandas
numpy
pickle
matplotlib.pyplot
seaborn
PIL
io
torch
transformers